proyecto_2/requirement.txt
matplotlib==3.8.2
pandas==2.1.4
pillow==10.2.0
seaborn==0.13.1
Unidecode==1.3.7